Implemented requested changes
