+++ README.md
@@ -74,9 +74,9 @@ struct ContentView: View {
         @State private var visibility: TabBarVisibility = .visible
 
     var body: some View {
-        TabBar(selection: $selection, visibility: $visibility) {
+                TabBar(selection: $selection, visibility: $visibility) {
             HStack{
-                Button(action: {
+                Button {
                     switch visibility {
                     case .visible:
                         withAnimation {
@@ -87,7 +87,7 @@ struct ContentView: View {
                             visibility = .visible
                         }
                     }
-                }) {
+                } label: {
                     Text("Hide/Show TabBar")
                 }
             }
